requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-groq
langchain-community
faiss-cpu